pyproject.toml
[tool.poetry]
name = "mlbackendffej"
version = "0.1.0"
description = ""
authors = ["iyunakov <iyunakov@inbox.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.111.0"
sqlalchemy = {extras = ["asyncio"], version = "^2.0.30"}
asyncpg = "^0.29.0"
alembic = "^1.13.1"
uvicorn = {extras = ["standard"], version = "^0.30.1"}
httpx = "^0.27.0"
coverage = "^7.5.3"
pytest = "^8.2.2"
redis = "^5.0.5"
colorlog = "^6.8.2"
pydantic-settings = "^2.3.1"
email-validator = "^2.1.1"
celery = "^5.4.0"
flower = "^2.0.1"
requests = "^2.32.3"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
bcrypt = "4.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
